Add search-list support to socks proxy
This adds search list support as documented in resolv.conf(5) with
a fixed dot value of 1. All domain names with 1 or more dots will be
looked up first as is, and then via the domain search list. All domain
names with zero dots will be lookup up via the domain search list
first, and then as is.